# packages/proxy/src/openapi.yaml
openapi: 3.1.0
info:
  title: HAAI API
  description: |
    Management REST API and OpenAI-compatible inference endpoints for HAAI.

    **Management API** (`/api/v1/*`) requires admin authentication via session cookie or Bearer token.

    **Inference API** (`/v1/*`) requires an API key in the `Authorization: Bearer` header.
  version: 0.0.1
servers:
  - url: /
    description: Current host
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: haai_session
    bearerAuth:
      type: http
      scheme: bearer
    apiKeyAuth:
      type: http
      scheme: bearer
      description: Client API key (haai-sk-…)
...
paths:
...
  /api/v1/vmodels:
    get:
      tags: [Virtual Models]
      summary: List virtual models
      security: [{ sessionCookie: [] }, { bearerAuth: [] }]
      responses:
        "200":
          description: Virtual model list
    post:
      tags: [Virtual Models]
      summary: Create virtual model
      security: [{ sessionCookie: [] }, { bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [modelId, backends]
              properties:
                modelId: { type: string }
                displayName: { type: string }
                balancingStrategy:
                  type: string
                  enum: [session-pin, round-robin, weighted, least-connections, least-latency]
                streaming: { type: boolean }
                allowToolCalling: { type: boolean }
                backends:
                  type: array
                  items:
                    type: object
                    properties:
                      backendId: { type: string }
                      backendModelId: { type: string }
                      weight: { type: integer }
      responses:
        "201":
          description: Virtual model created
  /api/v1/vmodels/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string }
    get:
      tags: [Virtual Models]
      summary: Get virtual model
      security: [{ sessionCookie: [] }, { bearerAuth: [] }]
      responses:
        "200":
          description: Virtual model details
    patch:
      tags: [Virtual Models]
      summary: Update virtual model
      security: [{ sessionCookie: [] }, { bearerAuth: [] }]
      responses:
        "200":
          description: Updated virtual model
    delete:
      tags: [Virtual Models]
      summary: Delete virtual model
      security: [{ sessionCookie: [] }, { bearerAuth: [] }]
      responses:
        "204":
          description: Deleted
...